BUGG-119 Add missing setting for subscriptions
There is a configuration variable for Subscriptions module
that renders the form as a block. We need this variable for
Display Suite or the form won't show up in Guide Pages and
Images.

I also took the chance to rename the module to bgsubscriptions
so it stays consistent with other custom modules and avoids
potential hook colisions such as hook_subscriptions.