#52 Only allow moving images to the images tab.
Currently the move button appears on bgpage and bgimage nodes, but we only want
it to appear on the images tab of the taxon being moved to.

menu_get_object() returns null on a node/%/bgimages page, so I switched to
menu_get_item and used arg() to get the current nid.