#52 Set the correct parent for a tagged image being moved.
The new parent of the images being moved shouldn't be the same as the parent of
the node being moved to; the new first parent of the images being moved should
be the node being moved to.

This commit will still fail when moving to node 3 - that will be fixed in the
next commit.