#50 Provide a link to a full-sized image based on user role.
A user should be able to view a full size image if either:
* their user role has the 'view fullsize bgimages' permission, or
* they created the image.

As suggested at https://github.com/bugguide/bugguide/issues/50#issuecomment-736978382
this was implemented by modifying the image module's version of
hook_field_formatter_view.

Instead of relying on testing $display['settings']['image_link'] == 'file' as in
the image module, which (as far as I can tell) would require making image_link =
'file' the default setting, we instead test for an entity type of bgimage. That
way we can keep the default image_link setting of '' and make our decisions
about full-size viewing knowing that we're only dealing with a bgimage image -
plus a bgimage node should be the only place where it's possible to get a
full-size link of an image (I think), so we want that extra layer of protection
anyway.

bgimage_field_formatter_info was added since bgimage_field_formatter_view would
never be called without hook_field_formatter_info notifying that a formatter is
present in the bgimage module. In order for the formatter to be recognized and
called for default bgimage images, the 'type' of the default field display
setting must be set to the name of the formatter in
bgimage.features.field_instance.inc.