#249 Always base the series of a clone on the original image of the series.
It was being based on whichever image in the series was cloned from, i.e. whichever image the 'Add image of this individual' button was clicked on - so in effect if you cloned off of the second image in a series you were starting a brand new series (consisting of the second image and the new cloned image) which no longer included the first image in the original series.

For the record, no matter which image in the series the 'Add image of this individual' button is clicked on, the new cloned image is always added to the end of the series (not inserted). (If you want to reorder you have to unlink, add to clipboard in the order you want, and then relink.)