#298 Don't show 'Add to clipboard' for an image not belonging to you.
The patch here assumes the only two operations any user might use the clipboard for are a) linking images, and b) moving images.

With this check in place, technically we could probably remove some other code that does similar checks when it comes time to actually link or move, but we don't do that here (seems better to leave the extra checks).