#335 Check for a living being on the user register form.
I've tried to duplicate the current BugGuide behavior here, which as far as I can tell is:
* if form submission fails because the username or email is already taken, then give the user another try;
* otherwise redirect to the help page (which doesn't exist on bg7 yet), I guess to make it harder for bots to try again?

Unfortunately this means we have to try to check the error messages set by the system to detect why system validation failed (this will probably need more work if we ever translate to other languages).

The redirect will currently 404 since we don't have a /help page yet.

Note that we set the title on the 'Site' field using the placeholder attribute, following the practice of the other fields on this form, but then also set an invisible actual title because of https://github.com/bugguide/bugguide/blob/5030d564dd347b44422b0876fe8deb4f8e4f8d1f/includes/form.inc#L1452-L1462
"// Although discouraged, a #title is not mandatory for form elements. In
 // case there is no #title, we cannot set a form error message.
 // Instead of setting no #title, form constructors are encouraged to set
 // #title_display to 'invisible' to improve accessibility."