#352 Display a 'Home >> Recent' breadcrumb on the Recent page
Note that current_path() returns 'node' for the Home link bugguide.test/, the same as for bugguide.test/node, so we need to use request_path() (which returns '' for bugguide.test/) here instead. As far as I know that makes no difference when you pass the result to path_is_admin().