#205 Don't show tabs if there's only one
We never want to show the View tab on ID Request or Frass, and in other cases we don't want to show it if it's the only tab. This means no tabs are shown on ID Request or Frass for anyone, even editors and admin - I don't think there's any reason for editors to edit those pages under normal circumstances, but if there's a need one can always just type out the edit link by hand (if you do, the 'View' tab on the edit page links to /node/6, which is wrong).