#89 Display bglink card links as in the bglink teaser view
That's accomplished (as in the teaser view) by placing them in a <ul class="links inline"> for which we already have css defined.

The $handler->display->display_options['row_options']['default_field_elements'] = FALSE; setting is unchecking the 'Provide default field wrapper elements' in the 'Show' settings under FORMAT at admin/structure/views/view/links/edit; without that each of our field outputs gets a wrapper and then the css doesn't work (as defined).

The result here mimics what the teaser view adds by default, but differs slightly from current BugGuide in that current BugGuide doesn't show the 'add comment' link if there are already comments (on one hand that uses less space, which I like; on the other I think this is more consistent and logical, and it's what the teaser view does...).

Current BugGuide includes a 'Subscribe' link as well, but when I try to add a 'Content: Subscribe link' to the view it reports 'Broken/missing handler', so that will need to be addressed in a followup.