#89 Show the 'Add new comment' link on bglink cards as long as user is logged in
It seems that the comment module's behavior is to only show the 'Add new comment' button if there are already comments: https://github.com/bugguide/bugguide/blob/e57007cbea78396abc06c0a0ba15cd28f1414e10/modules/comment/comment.module#L674-L681
    "// Show the "post comment" link if the form is on another page, or
     // if there are existing comments that the link will skip past."

That doesn't make sense to me in our situation (and is actually the opposite of what current BugGuide does - current only shows the Add comment link if there _aren't_ any comments), so we're always showing the Add Comment button now, as the bglink teaser display does.

We could use a Rewrite result on the 'Add comment link' field to always display the 'Add new comment' text, but that doesn't offer access control, so we do it here instead; the other benefit is that doing it this way we still automatically get the 'Log in or register to post comments' text if the user isn't logged in.