Revert "#281 Honor an empty date when creating a cloned image."
This reverts commit f5e342eba9c5829e00644be2163ddfec8d6d60be.
When I created that commit I didn't realize the non-empty date was a result of a default date of 'today' set on the bgimage date field; we now have no default date (as on current BugGuide), so there's no need for that fix. Also, for reasons unkown to me, the commit being reverted never fixed the problem on beta, only locally, so I'm glad for another fix!