#390 Fix breadcrumbs for ID Request and Frass
There are two separate issues here:

1. The system breadcrumb for ID Request and Frass is set here: https://github.com/bugguide/bugguide/blob/4da7f607a002f34a1c0d2bcabc5c3f3d4aa8f6bd/sites/all/modules/custom/bg/bg.module#L497
Locally, from the repo, the parent field for the ID Request and Frass nodes is set, but equal to the empty string, so locally we were displaying a breadcrumb on those pages. On beta no system breadcrumb was being displayed (I'm not sure why). Since ID Request and Frass both already display their title as part of the page and display separate breadcrumb links as part of the images div, this commit removes the system breadcrumb for ID Request and Frass (by checking empty instead of !isset on the parent field in bg_create_taxonomic_breadcrumb).

2. When breadcrumbs are created for a path that includes ID Request and Frass, they need to use the bgimage link (instead of the usual info link) for the ID Request/Frass node since those nodes are special - doing that fixes the breadcrumb links that appear at the top of the image div for images displayed on ID Request and Frass. For those breadcrumbs bg_create_taxonomic_breadcrumb is being called for each image in ID Request or Frass here: https://github.com/bugguide/bugguide/blob/4da7f607a002f34a1c0d2bcabc5c3f3d4aa8f6bd/sites/all/modules/custom/bgimage/bgimage.module#L186-L190; since it's being called on an image that has a parent (the parent is ID Request or Frass), it technically gets around the 'do we have a parent' check in bg_create_taxonomic_breadcrumb, even though in spirit it probably shouldn't (since the check is meant to exclude the case where only one non-Home breadcrumb would be created).