#167,#242 Sort the images tab first by parent taxa nids, then by image nid
I think this replicates what current BugGuide does.

We're sorting first by parent nids, as a string of comma-separated parents - so sorted in dictionary order, where ',' comes before all digits - in ascending order (as on current BugGuide); and then by image entity_id = nid, which in solr is an integer, in descending order, so most recent images are displayed first.