#24,#414 Add the bgimage avg photo quality rating to Solr
We need it for filtering.

Note that we've converted the 'average' value from the rating fields - which is a string representation of a float (an average of ints) - into an int by applying intval, which truncates the float - if you try to send Solr a string representation of a float that doesn't read as an int, you get an error.

(I added the new field to bg.apachesolr_enfironments.inc, though I don't know if it's necessary or not - it just looked like all the other fields were listed in that file.)