#415 Provide indications when the user can't browse any lower.
Note that we added a span around the $title_markup in the $leaf_taxon == TRUE case - without the span, the display: flex of bulma's card-header-title class was causing issues when $title_markup included an <i> section, e.g. $title_markup = 'honey bee <i>Apis melifera</i>'. No space was displayed after 'bee', and when the title div got narrow, the italics portion would shrink by stacking its text instead of keeping it inline and wrapping. [I guess 'honey bee ' and '<i>Apis melifera</i>' become separate flex items (you can see that in the inspector) and then the space at the end of 'honey bee ' gets collapsed.]

Wrapping the title in a <span> fixes both of those issues - presumably this issue was never apparent before because prior to this commit the title text was always wrapped in an <a>, which had the same grouping effect as the span.