#401 Hide the 'Extract image data' button unless an image is selected.
When you first edit an image node that already has an image, or after you've uploaded an image the user selected, the upload/extract button doesn't exist, instead there's a 'Remove' button.

Otherwise the upload button exists and this code using the '#states' value seems to do what we want by only making the button visible when the value of the file input is not '', i.e. visible only when a file has been selected.

[The #states notation '!value' => '' means "value != ''".]