#427 Display a user's vote for fivestar fields
Unfortunately I feel like each of the available options has issues. I think the option chosen here is the best for logged in users; the downside is for anonymous users they just see 5 empty stars that they can't click on and don't know what they're there for, though they do get the text reporting the number of votes and the average rating. (I considered hiding the stars for anonymous users, but then they wouldn't know what the average rating was out of (there are 5 stars), so then I added in 'out of 5' text with CSS ::after content, but that seems like something you're-not-really-supposed-to-do?)

The 'Value to display as stars' value of 'User's vote if available, average otherwise' would be more attractive to me if the style of the user's vote was different than the style of the average vote (so you could tell which was being shown! For me the whole point of this bug is if I visit a not-recent image, I want to be able to tell if I've voted on it already or not, and if I have what my vote was.). Also when you're logged in the number of stars displayed for the average is always rounded up to the next largest integer, so an average of 3.3 would be displayed as 4 - in my mind that makes it impossible to know if you voted 4 or if it's just showing the average 3.3 as 4. (When you're not logged in it actually displays three and a third filled in stars.)