#427 Add a new fivestar formatter
The formatter adds a new option for displaying stars: if the user is logged in it displays the number of stars they've voted for (0 if they haven't voted), while if they're logged out it displays the average number of stars.

The various formatter functions set their display type to 'fivestar_formatter_default' prior to calling fivestar since fivestar's formatter checks for that.

[The existing option "User's vote if available, average otherwise" shows the average vote if a user is logged in and hasn't voted yet - unfortunately there's no way for a user to tell if they haven't voted and the average is being shown or if they have voted and their vote just happens to be the same as the average; an issue compounded by the fact that for logged in users, the average displayed on the stars (but not the text below) is always rounded up, so if the average rating is 3.3, 4 stars will be filled in for the average rating.]