#389 Allow users to enter 2-digit years for bgimages
The date popup accepts dates of the form mm-dd-yyyy, but we'd like to allow users to enter 2-digit years as well, as they can on current BugGuide.  On current BugGuide the rule for transforming a 2-digit year into a 4-digit year is that followed by php's date formats, see the first Note at https://www.php.net/manual/en/datetime.formats.date.php on the y and yy formats.

Since we don't allow dates in the future, I've altered that here so that any two-digit year less than the current two-digit year is interpreted as being in the 2000s, and otherwise (if the year entered is larger than the current year) it's interpreted as being in the 1900s. If that doesn't match the user's intent then they'll have to enter a four-digit year manually instead.

Implementation note:
Initially I tried to implement this by making the required changes through hooks in the date validation code, namely
hook_date_combo_pre_validate_alter() and hook_date_popup_pre_validate_alter(). That worked for the latter hook, but the former hook copies out the date that's going to be validated prior to calling the alter hook, so there was no way to make that hook work for us here. Instead I went with an #after_build function on the bgimage form - though the documentation I've read online suggests that's normally used when a form is built, it's also called on form submission, which is where I'm using it here: #after_build is called before any validation and allows us to alter the $form_state after it's been populated with 'input' and 'values' arrays, precisely what we need here.