#389 Add the date format being used by the field_bgimage_date input
This adds the date format 'n/j/Y' to the admin page admin/config/regional/date-time/formats
which makes it available as an option at admin/structure/types/manage/bgimage/fields/field_bgimage_date
in the 'Date entry options' drowpdown in the 'More settings and values' section [it's a little confusing because that dropdown doesn't show the n/j/Y date format (or any other format) as an option, it just shows a date formatted in that format, e.g. 12/13/2021, as an option]. We've already specified exactly that format for that setting in the bgimage feature; now that we've actually made it an option in the 'Date entry options' dropdown, it's now being correctly selected. (Formerly that dropdown showed "Short date format ...", presumably because the actual date format wasn't available and it was just showing the first option in the dropdown - the actual date input display has always been correct from the feature (n/j/Y), it simply wasn't showing up correctly in that dropdown in the admin menu).

With more detail:
On the Edit page for field_bgimage_date, admin/structure/types/manage/bgimage/fields/field_bgimage_date, the 'Date entry options' $options come from date_popup_formats():
https://github.com/bugguide/bugguide/blob/58cf8a056e0fe4b9ab348bf510557b104c926884/sites/all/modules/date/date_admin.inc#L329

The date_popup_formats() are a merge of the formats associated with the 'short' type and those associated with the 'custom' type:
https://github.com/bugguide/bugguide/blob/58cf8a056e0fe4b9ab348bf510557b104c926884/sites/all/modules/date/date_popup/date_popup.module#L693-L700

So if we want 'n/j/Y' to appear in the 'Date entry options' dropdown, it either needs to be associated to the 'short' or 'custom' type.

Only those formats whose type is 'custom' appear at admin/config/regional/date-time/formats:
https://github.com/bugguide/bugguide/blob/58cf8a056e0fe4b9ab348bf510557b104c926884/modules/system/system.admin.inc#L2870-L2887
I feel like the format showing up in the admin menu is a good thing (since I've gone to the trouble of defining it in hook_date_formats()), so I've chosen the 'custom' type for the 'n/j/Y' format.

With that choice, the 'n/j/Y' format appears at admin/config/regional/date-time/formats, and since we've set the 'input_format' for field_bgimage_date to be 'n/j/Y' in the bgimage feature: https://github.com/bugguide/bugguide/blob/58cf8a056e0fe4b9ab348bf510557b104c926884/sites/all/modules/custom/bgimage/bgimage.features.field_instance.inc#L306 and that format is now an option in the 'Date entry options' dropdown, that format is now correctly selected (instead of the first option being shown as selected, which was what was formerly happening, presumably because the option marked as selected, 'n/j/Y', didn't formerly exist in the options list).