#402 Provide feedback on what EXIF data was extracted on image submission
This was written assuming there will eventually be feedback to provide on other fields in addition to date.

The item to which to prepend the feedback (.form-item-field-bgimage-image-und-0) was chosen so that it would be overwritten if/when the user clicks the Remove button to choose a different image (the outer image field class '.field-name-field-bgimage-image' doesn't have that property).