#451 Validate that bgpage Representative Image values are all of type bgimage.
We break after each nid that fails since form_set_error only displays the first error message set on a given form element (so I'm choosing to break after each error instead of concatenating all error messages for that field).

The checks here may be stricter than current BugGuide. For example '3, 4' with a space after the comma is not allowed now (not sure if it is currently). If it is stricter then the first person to edit that guide page on BugGuide 2 will wind up needing to fix it.